Rename Subtable.js to MaterialsTable.js
